Remove sfImageMagickAdapter references
I needed to remove it from three places:

1. From the digital object derivatives setting action. imagick handles
   PDF page extraction, not pdfinfo any more. So I changed the error
   message to reflect that.
2. From the QubitDigitalObject class. I was using the
   sfImageMagickAdapter as a fallback for sfImagickAdapter, so I removed
   those fallback cases.
3. From the lime unit tests in the sfThumbnailPlugin.

The unit tests do run successfully in the sfThumbnailPlugin class, but
it's tedious to run them.

In case you want to run them, you need to run these commands in the
Docker container to install gd.

    apk add libpng-dev libjpeg-turbo-dev
    docker-php-ext-configure gd --with-jpeg
    docker-php-ext-install gd

This is because the tests *require* you to have gd installed, and gd is
not installed in the base AtoM image.

And then to run the tests, you invoke the test file directly:

    php plugins/sfThumbnailPlugin/test/unit/sfThumbnailTest.php

You'll see this when the tests run:

    # Looks like everything went fine.
    # Looks like everything went fine.